RISE_OML/ai2_newVersion/requirements.txt
fastapi
uvicorn
yfinance

google-cloud-storage